Safari-611.1.10/LayoutTests/http/tests/security/no-indexeddb-from-sandbox-expected.txt

ALERT: PASS: db.open() threw a SECURITY_ERR!
